% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/autoroxy.R
\name{autoroxy}
\alias{autoroxy}
\title{Generate roxygen2 documentation during package build}
\usage{
autoroxy()
}
\description{
Add a call to this function to any file in your \code{R} directory.
Use explicit qualification (like \code{autoroxy::autoroxy()}) to avoid
importing the package and its dependencies.
}

