% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoroxy.R
\name{autoroxy}
\alias{autoroxy}
\title{Generate roxygen2 documentation during installation from source}
\usage{
autoroxy()
}
\description{
Add the body of this function
to any file in your \code{R} directory.
(This is done automatically by \code{\link{rox_off}}.)
}

